+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - BREAKING: `cleanup_on_success` & `cleanup_on_failure` replace `close_on_job_exit`.
 - `cleanup_on_success` & `cleanup_on_failure` control not only closing the terminal window but also cleaning up the terminal instance.
 - Fixed vim-ui-select picker not working as expected.
+- Fix previewing non-active sticky terminals
 
 ## [0.2.0] - 2025-07-31